end edit event

+++ admin_event/event_programme/report.php
@@ -3,15 +3,13 @@
         <div class="bread">
             <nav style="--bs-breadcrumb-divider: '>';" aria-label="breadcrumb">
                 <ol class="breadcrumb">
-                    <li class="breadcrumb-item"> <i class="fa fa-heart"></i> <a
-                            href="main.php?dir=dashboard&page=dashboard"> الاحداث </a> <i
-                            class="fa fa-chevron-left"></i> </li>
-                    <li class="breadcrumb-item active" aria-current="page"> برنامج الاحداث  </li>
+                    <li class="breadcrumb-item"> <i class="fa fa-heart"></i> <a href="main.php?dir=dashboard&page=dashboard"> الاحداث </a> <i class="fa fa-chevron-left"></i> </li>
+                    <li class="breadcrumb-item active" aria-current="page"> برنامج الاحداث </li>
                 </ol>
             </nav>
         </div>
         <div class="title text-right">
-            <h6> <i class="fa fa-search"></i>   برنامج الاحداث  </h6>
+            <h6> <i class="fa fa-search"></i> برنامج الاحداث </h6>
         </div>
         <!-- Content Row -->
         <div class="table-responsive">
@@ -28,19 +26,17 @@
                         $stmt->execute();
                         $alltype = $stmt->fetchAll();
                         foreach ($alltype as $type) { ?> <tr>
-                        <td><?php echo $type['prog_name']; ?> </td>
-                        <td><?php echo $type['event_page']; ?> </td>
-                        <td>
-                            <a class=" btn btn-success"
-                                href="main.php?dir=event_programme&page=edit&prog_id=<?php echo $type['prog_id']; ?> ">
-                                <i class="fa fa-edit"></i>
-                            </a>
-                            <a class="confirm btn btn-danger"
-                                href="main.php?dir=event_programme&page=delete&prog_id=<?php echo $type['prog_id']; ?> ">
-                                <i class="fa fa-trash"></i>
-                            </a>
-                        </td>
-                    </tr> <?php }
+                            <td><?php echo $type['prog_name']; ?> </td>
+                            <td><?php echo $type['event_page']; ?> </td>
+                            <td>
+                                <a class=" btn btn-success btn-sm" href="main.php?dir=event_programme&page=edit&prog_id=<?php echo $type['prog_id']; ?> ">
+                                    <i class="fa fa-edit"></i>
+                                </a>
+                                <a class="confirm btn btn-danger btn-sm" href="main.php?dir=event_programme&page=delete&prog_id=<?php echo $type['prog_id']; ?> ">
+                                    <i class="fa fa-trash"></i>
+                                </a>
+                            </td>
+                        </tr> <?php }
                                 ?> </tbody>
             </table>
         </div>
